inviteduser não precisa de plano

+++ src/lib/actions/socialMedias/social.delete.action.tsx
@@ -6,13 +6,6 @@ import { hasSubscription } from "@/lib/helpers/billing";
 import { SocialMediaTypes } from "@prisma/client";
 
 export const deleteSocial = async (type: SocialMediaTypes) => {
-  const hasSub = await hasSubscription();
-
-  if (!hasSub) {
-    return {
-      error: `Você precisa assinar um plano para realizar este comando.`,
-    };
-  }
 
   const session = await getSession();
   if (!session?.user.id) {
@@ -20,6 +13,20 @@ export const deleteSocial = async (type: SocialMediaTypes) => {
       error: "Not authenticated",
     };
   }
+
+   const user = await prisma.user.findUnique({
+     where: {
+       id: session.user.id,
+     },
+   });
+
+   const hasSub = await hasSubscription();
+
+   if (user?.role === "politician" && !hasSub) {
+     return {
+       error: `Você precisa assinar um plano para realizar este comando.`,
+     };
+   }
 
   const response = await prisma.socialMedia.delete({
     where: {
